+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Implemented `is_forestal` trait. Entities with this trait gain defense (+3) during combat when positioned on forest terrain. Also added forest terrain and forestal entities support to the CLI visualizer.
 
 ## [Unreleased]
+## [2026-08-31]
+- Implemented `is_autumn_dweller` trait. Entities with this trait treat the autumn season as a shelter, gaining increased defense and energy recovery.
 - Implemented `is_storm_dweller` trait. Entities with this trait treat `storm` conditions as a shelter, gaining increased defense and energy recovery.
 - Implemented `is_ice_dweller` trait. Entities with this trait treat `ice` terrain as a shelter, gaining increased defense and energy recovery.
 - Implemented `is_mud_dweller` trait. Entities with this trait treat `mud` terrain as a shelter, gaining increased defense and energy recovery.